Add flight query dialog and reusable list statistics
